Move Identity shape inference into an identity submodule
Move the Identity operator, plus its test, from ops.rs into a new
ops/identity.rs submodule, matching how this operator is grouped in
src/ops.rs.
